modify some code to test handmade sample cases and add corner cases
