+++ tests/utils.py
@@ -4,6 +4,14 @@ import pathlib
 import subprocess
 import sys
 import tempfile
+from typing import *
+
+import bs4
+
+import onlinejudge._implementation.logging as log
+import onlinejudge._implementation.testcase_zipper
+import onlinejudge._implementation.utils as utils
+from onlinejudge.type import *
 
 
 @contextlib.contextmanager
@@ -89,3 +97,16 @@ def is_logged_in(service, memo={}):
         proc = run(['login', '--check', url])
         memo[url] = proc.returncode == 0
     return memo[url]
+
+
+def get_handmade_sample_cases(self, *, html: str) -> List[TestCase]:
+    # parse
+    soup = bs4.BeautifulSoup(html, utils.html_parser)
+    samples = onlinejudge._implementation.testcase_zipper.SampleZipper()
+    for pre in soup.select('.sample pre'):
+        log.debug('pre %s', str(pre))
+        it = self._parse_sample_tag(pre)
+        if it is not None:
+            data, name = it
+            samples.add(data.encode(), name)
+    return samples.get()
